Viz layer

+++ models/base/csv_uploads/combined_sales.sql
@@ -7,9 +7,9 @@ SELECT
   Marketplace AS marketplace,
   Royalty_Type AS royalty_type,
   Transaction_Type AS transaction_type,
-  IFNULL(CAST(Units_Sold AS INT64),0) AS units_sold,
-  IFNULL(CAST(Units_Refunded AS INT64),0) AS units_refunded,
-  IFNULL(CAST(Net_Units_Sold AS INT64),0) AS net_units_sold,
+  IFNULL(CAST(Units_Sold AS numeric),0) AS units_sold,
+  IFNULL(CAST(Units_Refunded AS numeric),0) AS units_refunded,
+  IFNULL(CAST(Net_Units_Sold AS numeric),0) AS net_units_sold,
   IFNULL(CAST(Avg__List_Price_without_tax AS numeric),0) AS avg_list_price_without_tax,
   IFNULL(CAST(Avg__Offer_Price_without_tax AS numeric),0) AS avg_offer_price_without_tax,
   CASE WHEN Avg__Delivery_Manufacturing_cost = 'N/A' THEN 0
